Initial commit
Initial commit.
The test passes but is not yet implemented in cmake i.e. it has to be
built manually.
The Doxygen documentation is implemented altough some changes should
probably be done to the Doxyfile to make it a bit more usable. Also the
tests directory should be omitted from the documentation.