Implemented meshing and simulation modules
Implemented the meshing module.
Still a bit rough but works fairly well. Will probably have some issues
with extreme cases with low pitch/high stagger so a backup strategy for
creating the section geometry needs to be implemented at some point.

Implemented the simulation module.
Very rough, but works. Missing postprocessing, but will work on this in
conjunction with the main design loop.
Currently sources the case template from raw files, these should be
moved to headers, so the source location variable doesn't have to be
loaded into the environment. Doing this next.

Implemented basic output for RegistryObjects and the Registry.
Should probably add something that orders whats printed in some kind of
order because the output is a bit hard to read.
Should also add merging slave and master Registries or some simple way
that a master Registry can access slave Registry data.

Moving on to the rudimentary parser so runtime testing can be done more
easily.
Postprocessing and the main design loop come after this.

Still missing of unit tests :(