Reimplemented Blade
Reimplemented the 'Blade' class.
Currently requires that there be atleast 3 airfoils for it to be built.
Will adjust this so that it can be built with 1 airfoil so meanline
analysis can be performed.

[BUG]:
Theres currently a bug with how the Blade is being built in that it
generates 2 extra surfaces (pressure/suction) along with the volume we
want. This is because gmsh's loft doesn't allow the sharing of curves
between lofts, so when we're creating the Blade volume we have to stitch
the pressure and suction surfaces and this creates new surfaces which
share topological curves on their boundaries.
We can fix this by building from sections instead of vertical splines,
but we would have to implement our own way of interpolating between
profiles (eg. by implementing our own nurbs/b-splines).
We're leaving it as is but this should be fixed.

Changed the pressure requirement from total to static and adjusted the
design equations to reflect this.

Added automatic computing of the vortex distribution coefficient such
that the required work (based on the requested static pressure) is
satisfied. Specfication by user definition is also allowed.

Added some geometry classes and operators.
Added some math classes and utilities.
Moved the 'turbo_case_template' files to headers.

Fixed a bug with geometry removal, should always remove properly now.
Fixed a bug with model removal and tracking.

We're missing unit tests.