Reimplemented design loops
The design algo will now first align the stagnation point with the
leading edge, and then first adjust the camber to try to satisfy the
required pressure. After each adjustment it will realign the stagnation
point.
If the max camber adjustment is reached without satisfying the target
pressure, the algo will then adjust the stagger angle to again try to
satisfy the pressure.
Both the camber and the stagger angle adjustments handle
over/undershooting the pressure and will interpolate the adjustment step
so that the target pressure is satisfied within the specified tolerance.

As an aside, the wrapped profile .csv export in the Blade.cpp is
currently commented out, but it should actually be implemented (maybe
like a dedicated function or add it to the main src file?).

Also the way the blade is being built actually causes considerable
issues when the .step has to be used later on (eg. in some other cad
software). This should be what we work on next along with testing how
different design cases are handled in general.