Minor changes and bugfixes
Fixed a bug with how Polyline interpolates it's values. Although the
Polyline class should be reworked because it's probably still fairly
boned. Currently scalar values are interpolated 'correctly' but it's
probably fairly unintuitive. How it behaves when it's used as a
geometric entity is a mistery.

The stagnation point alignment algo will return the data from the last
sim it runs. Also the profile will be written to the current sim case
directory when Airfoil::simulate() is called if the 'writeProfile' flag
is set instead of it having to be called independantly.

Minor changes to default foam simulation settings.

AirfoilEfficiency can now be greater than 1. This is so that we don't
throw if our sims don't converge properly and we get garbage values
because there's a good possibility that we'll manage to recover
(converge) during the next sim (design loop iteration).